+++ packages/showcases/react/.storybook/main.js
@@ -8,7 +8,9 @@ module.exports = {
     '@storybook/addon-a11y',
     '@whitespace/storybook-addon-html',
     'storybook-addon-designs',
-    'storybook-mobile',
-    'storybook-addon-outline',
+    'storybook-mobile'
   ],
+  features: {
+    postcss: false,
+  },
 };