+++ packages/showcases/svelte/.storybook/main.js
@@ -1,14 +1,16 @@
 module.exports = {
-  'stories': [
+  stories: [
     '../stories/**/*.stories.{js,jsx,ts,tsx,mdx,svelte}'
   ],
-  'addons': [
+  addons: [
     '@storybook/addon-svelte-csf',
     '@storybook/addon-essentials',
     '@storybook/addon-a11y',
     '@whitespace/storybook-addon-html',
     'storybook-addon-designs',
-    'storybook-mobile',
-    'storybook-addon-outline',
-  ]
+    'storybook-mobile'
+  ],
+  features: {
+    postcss: false,
+  },
 }