+++ packages/showcases/svelte/stories/components/VtmnPopover/VtmnPopover.stories.svelte
@@ -1,13 +1,13 @@
 <script>
   import { Meta, Template, Story } from '@storybook/addon-svelte-csf';
-  import VtmnPopover, {
-    POSITION,
-  } from '@vtmn/svelte/src/components/VtmnPopover.svelte';
-  import VtmnButton from '@vtmn/svelte/src/components/VtmnButton.svelte';
+  import { VtmnPopover, VtmnLink, VTMN_POPOVER_POSITION } from '@vtmn/svelte';
+  import { parameters } from '@vtmn/showcase-core/csf/components/VtmnPopover.csf';
 
   const popoverArgs = {
+    id: 'My popover',
     title: 'This is the title of the popover',
-    body: 'A popover can appear when users click or focus an element to provide his description in details, and often also with a link or button inside. Popovers will automatically close when clicking outside the popover bounds.',
+    body: 'A popover can appear when users click or focus an element.',
+    position: VTMN_POPOVER_POSITION.BOTTOM,
   };
 </script>
 
@@ -17,54 +17,20 @@
   argTypes={{
     position: {
       control: 'select',
-      options: Object.keys(POSITION),
-      mapping: POSITION,
+      options: Object.values(VTMN_POPOVER_POSITION),
     },
-    title: { control: 'text' },
-    body: { control: 'text' },
   }}
+  {parameters}
 />
 
 <Template let:args>
-  <div class="story-container">
-    <VtmnPopover position={args.position}>
+  <div>
+    <VtmnPopover position={args.position} id={args.id}>
       <svelte:fragment slot="title">{args.title}</svelte:fragment>
       <svelte:fragment slot="body">{args.body}</svelte:fragment>
-      <VtmnButton slot="placeholder">{args.position}</VtmnButton>
+      <VtmnLink slot="placeholder">Popover on the {args.position}</VtmnLink>
     </VtmnPopover>
   </div>
 </Template>
 
-<Story name="All positions">
-  <div class="story-container">
-    {#each Object.values(POSITION) as position}
-      <VtmnPopover {position}>
-        <svelte:fragment slot="title">{popoverArgs.title}</svelte:fragment>
-        <svelte:fragment slot="body">{popoverArgs.body}</svelte:fragment>
-        <VtmnButton slot="placeholder">{position}</VtmnButton>
-      </VtmnPopover>
-    {/each}
-  </div>
-</Story>
-
-<Story
-  name="With Controls"
-  args={{
-    position: POSITION.BOTTOM,
-    ...popoverArgs,
-  }}
-/>
-
-<style>
-  .story-container {
-    font-size: 1rem;
-
-    width: 100%;
-    height: 100vh;
-    display: flex;
-    align-items: center;
-    justify-content: center;
-
-    gap: 1rem;
-  }
-</style>
+<Story name="Overview" args={popoverArgs} />