+++ packages/sources/svelte/src/components/VtmnPopover.svelte
@@ -1,32 +1,19 @@
-<script context="module">
-  /** @enum */
-  export const POSITION = {
-    TOP_LEFT: 'top-left',
-    TOP: 'top',
-    TOP_RIGHT: 'top-right',
-    RIGHT: 'right',
-    BOTTOM_RIGHT: 'bottom-right',
-    BOTTOM: 'bottom',
-    BOTTOM_LEFT: 'bottom-left',
-    LEFT: 'left',
-  };
-
-  /** @type {POSITION} */
-  export const DEFAULT_POSITION = POSITION.BOTTOM;
-</script>
-
 <script>
   import '@vtmn/css-popover';
   import { cn } from '../utils/classnames';
+  import { VTMN_POPOVER_POSITION } from '../utils/enums';
 
-  /** @type {string} - An unique id */
-  export let identifier = '';
-  /** @type {POSITION} */
-  export let position = DEFAULT_POSITION;
+  /**
+   * ID of the popover
+   * @type {string}
+   */
+  export let id;
+  /** @type {VTMN_POPOVER_POSITION} */
+  export let position = VTMN_POPOVER_POSITION.BOTTOM;
 
   let className;
   /**
-   * @type {string} A custom class to apply to the component.
+   * @type {string} Custom classes to apply to the component.
    */
   export { className as class };
 
@@ -36,13 +23,13 @@
 <div
   class={componentClass}
   data-position={position}
-  aria-describedby={identifier}
+  aria-describedby={id}
   tabindex="0"
   {...$$restProps}
 >
   <slot name="placeholder" />
 
-  <div id={identifier} role="tooltip">
+  <div {id} role="tooltip">
     <p class="vtmn-popover_title"><slot name="title" /></p>
     <p class="vtmn-popover_text"><slot name="body" /></p>
   </div>