+++ packages/sources/svelte/src/index.js
@@ -1,1 +1,4 @@
+import '@vtmn/css-design-tokens';
+
 export * from './components';
+export * from './utils';