+++ packages/sources/css/src/components/tooltip/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.14](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tooltip@0.3.13...@vtmn/css-tooltip@0.3.14) (2021-10-20)
+
+**Note:** Version bump only for package @vtmn/css-tooltip
+
+
+
+
+
 ## [0.3.13](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-tooltip@0.3.12...@vtmn/css-tooltip@0.3.13) (2021-10-19)