+++ packages/sources/css/src/design-tokens/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.7](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-design-tokens@0.8.6...@vtmn/css-design-tokens@0.8.7) (2021-10-20)
+
+**Note:** Version bump only for package @vtmn/css-design-tokens
+
+
+
+
+
 ## [0.8.6](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-design-tokens@0.8.5...@vtmn/css-design-tokens@0.8.6) (2021-10-19)