+++ packages/sources/web-components/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.16.2](https://github.com/Decathlon/vitamin-web/compare/@vtmn/web-components@0.16.1...@vtmn/web-components@0.16.2) (2021-10-20)
+
+**Note:** Version bump only for package @vtmn/web-components
+
+
+
+
+
 ## [0.16.1](https://github.com/Decathlon/vitamin-web/compare/@vtmn/web-components@0.16.0...@vtmn/web-components@0.16.1) (2021-10-19)
 
 **Note:** Version bump only for package @vtmn/web-components